change `tryGet()` API misuse to `Defect`
Assuming a `Result[T, E]` with `E: Exception`, `tryGet` has the nasty
side effect of triggering `{.raises: [E, ResultError[void]].}`. This is
because it tries to play nice when `tryGet()` is called on a `Result`
that has not yet been initialized.

That is not in line with how Nim typically operates. Accessing items
that are not initialized may yield to SIGSEGV in many situations.

```nim
proc f() =
  var x: ref CatchableError
  raise x

f()
```

`tryGet()` semantics are well-defined for an initialized `Result`.
- If a value has been set, it is returned.
- If an error has been set, it is raised.

Calling `tryGet()` on un-initialized data is not a typical usecase
and could instead become a `ResultDefect`. This way, it is not required
anymore to `except + discard` the `ResultError[void]` for hypothetical
API misuse, or to alternatively pollute the callstack with `{.raises.}`.

The alternatiev would be to `raise E(msg: ...)` instead, but that one
may trigger unexpected behaviour as it hides the API misuse error.