`lent` support
`lent` allows returning a (hidden) pointer to a member thus making
Result slightly more efficient when working with types that are
expensive to copy.

`lent` is enabled by default only on recent versions as Nim is known to
generate invalid code in earlier releases - how far back `lent` works
reliably depends on where Result appears but versions as recent as
1.6.14 have been known to have issues.

As such, `lent` support is conservatively enabled only for 2.0.8+ but
can be enabled for earlier versoins as well via a compile-time define.